App is a container, not component
