+++ app/scripts/main-react.js
@@ -2,7 +2,7 @@ var React = require('react');
 var ReactDOM = require('react-dom');
 
 var Provider = require('react-redux').Provider;
-var App = require('./components/App');
+var App = require('./containers/App');
 var configureStore = require('./store');
 
 ReactDOM.render(
